requirements.txt
tensorflow>=2.10
flask
pillow
requests
numpy
python-dotenv
scipy